Move `io.element.android.tests.uitests.ScreenshotTest` to `ui.T` to generate much shorter name for screenshot files.
