+++ tests/uitests/src/test/kotlin/ui/ColorTestPreview.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.element.android.tests.uitests
+package ui
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
